gcp/requirements.txt
google-cloud-storage
tensorflow
numpy
Pillow